STT TTS & README update
