+++ chatbot_srv/.idea/chatbot_srv.iml
@@ -2,7 +2,7 @@
 <module type="PYTHON_MODULE" version="4">
   <component name="NewModuleRootManager">
     <content url="file://$MODULE_DIR$" />
-    <orderEntry type="jdk" jdkName="Remote Python 3.12.2 (sftp://bcml1@223.194.46.206:10001/home/bcml1/miniconda/envs/jaeyong/bin/python)" jdkType="Python SDK" />
+    <orderEntry type="jdk" jdkName="Python 3.11 (chatbot) (3)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
 </module>
