+++ chatbot_srv/.idea/misc.xml
@@ -1,4 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="ProjectRootManager" version="2" project-jdk-name="Remote Python 3.12.2 (sftp://bcml1@223.194.46.206:10001/home/bcml1/miniconda/envs/jaeyong/bin/python)" project-jdk-type="Python SDK" />
+  <component name="Black">
+    <option name="sdkName" value="Python 3.11 (chatbot) (3)" />
+  </component>
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.11 (chatbot) (3)" project-jdk-type="Python SDK" />
 </project>
